Implement base network available checking with auth

+++ app/src/main/java/com/tarasovvp/smartblocker/repository/AuthRepository.kt
@@ -1,13 +1,10 @@
 package com.tarasovvp.smartblocker.repository
 
-import android.content.Intent
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
-import com.tarasovvp.smartblocker.R
 import com.tarasovvp.smartblocker.SmartBlockerApp
-import com.tarasovvp.smartblocker.constants.Constants.EXCEPTION
-import com.tarasovvp.smartblocker.extensions.isNotTrue
+import com.tarasovvp.smartblocker.extensions.isTrue
 import com.tarasovvp.smartblocker.extensions.sendExceptionBroadCast
 
 object AuthRepository {
@@ -16,6 +13,7 @@ object AuthRepository {
     private val currentUser = FirebaseAuth.getInstance().currentUser
 
     fun sendPasswordResetEmail(email: String, result: () -> Unit) {
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         auth?.sendPasswordResetEmail(email)
             ?.addOnCompleteListener { task ->
                 if (task.isSuccessful) {
@@ -27,6 +25,7 @@ object AuthRepository {
     }
 
     fun signInWithEmailAndPassword(email: String, password: String, result: () -> Unit) {
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         auth?.signInWithEmailAndPassword(email, password)
             ?.addOnCompleteListener { task ->
                 if (task.isSuccessful) {
@@ -39,6 +38,7 @@ object AuthRepository {
     }
 
     fun signInWithGoogle(idToken: String, result: () -> Unit) {
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         val credential = GoogleAuthProvider.getCredential(idToken, null)
         auth?.signInWithCredential(credential)
             ?.addOnCompleteListener { task ->
@@ -55,6 +55,7 @@ object AuthRepository {
         password: String,
         result: () -> Unit,
     ) {
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         auth?.createUserWithEmailAndPassword(email, password)
             ?.addOnCompleteListener { createUserTask ->
                 if (createUserTask.isSuccessful) {
@@ -66,6 +67,7 @@ object AuthRepository {
     }
 
     fun changePassword(currentPassword: String, newPassword: String, result: () -> Unit) {
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         val user = FirebaseAuth.getInstance().currentUser
         val credential = EmailAuthProvider.getCredential(user?.email.orEmpty(), currentPassword)
         user?.reauthenticateAndRetrieveData(credential)
@@ -75,7 +77,8 @@ object AuthRepository {
                         if (passwordTask.isSuccessful) {
                             result.invoke()
                         } else {
-                            passwordTask.exception?.localizedMessage.orEmpty().sendExceptionBroadCast()
+                            passwordTask.exception?.localizedMessage.orEmpty()
+                                .sendExceptionBroadCast()
                         }
                     }
                 } else {
@@ -85,7 +88,7 @@ object AuthRepository {
     }
 
     fun deleteUser(result: () -> Unit) {
-        SmartBlockerApp.instance?.checkNetworkAvailable()
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         currentUser?.delete()?.addOnCompleteListener { task ->
             if (task.isSuccessful) {
                 result.invoke()
@@ -96,7 +99,7 @@ object AuthRepository {
     }
 
     fun signOut(result: () -> Unit) {
-        SmartBlockerApp.instance?.checkNetworkAvailable()
+        if (SmartBlockerApp.instance?.checkNetworkAvailable().isTrue()) return
         SmartBlockerApp.instance?.googleSignInClient?.signOut()
         auth?.signOut()
         result.invoke()
